splitting JSON implementation into several modules
