+++ ideas.cabal
@@ -161,6 +161,8 @@ Library
     Ideas.Text.HTML.Templates
     Ideas.Text.HTML.W3CSS
     Ideas.Text.JSON
+    Ideas.Text.JSON.Builder
+    Ideas.Text.JSON.Data
     Ideas.Text.JSON.Decoder
     Ideas.Text.Latex
     Ideas.Text.MathML
